Fix logic for c_safe_identifier to extension types
"visibility" in this case means "visibility to Python". It's
where the class comes from that matters to whether we mangle the
name.

It's slightly debatable what we should be doing for api or public
classes, but it's probably reasonable to treat them as the user's
responsibility to get them right.