Fix incorrect code generation for method override (GH-6705)
https://github.com/cython/cython/pull/3235 subtly changed the behavior of `declare`
to return a reference to an already declared `Entry` in the case of overriding a parent method.

This caused `cfunc_entries` to have two entries pointing to the same
`Entry` object, whereas `cfunc_entries` is expected to have two
different `Entry` objects, one for the parent method that is being
inherited, and one of the new override from the child class.

This PR fixes this bug by copying `Entry` instead of returning a reference.

Closes https://github.com/cython/cython/issues/6704