Some profiling/monitoring code object changes (#6703)
1. Never reuse frames for CYTHON_USE_MODULE_STATE (because the
assumption is that we may be re-using module state for a different
interpreter).
2. Also don't reuse code objects for CYTHON_USE_MODULE_STATE.
3. Pre-monitoring, if we're not re-using code objects, and have to
create a code object with __Pyx_createFrameCodeObject, then clean it up
after we've done with it (I'm not sure what circumstances we have to
create it though).
4. With sys-monitoring, always treat the code object as an owned
reference, since it's never static and so will leak if we have to create
it with __Pyx_createFrameCodeObject.

Related to #6664 (although noting free-threading actually looks OK since
it always uses sys-monitoring and so never caches static frames or code
objects variables).