Avoid exception normalisation of unused exception instances during "except" handling (GH-6602)
In trivial cases that "obviously" don't access the caught exception, we can avoid getting and normalising it completely. We already do this for `except …: pass`, but we could do it in a few other cases as well.

Note that we cannot allow any operations that might trigger user code to be executed since it might be looking at `sys.exc_info`. In the optimised case, we don't fill that any more.

Closes https://github.com/cython/cython/issues/6601