Make CachedCFunction mechanism thread-safe (#6669)
Use an atomic flag indicating the initialized-status. If in the
intermediate "initializing" state then be lock-free and operate on a
non-shared version instead.

On non-freethreading versions, handle the edge-case where another thread
manages to initialize it while we're working (and so a reference count
is lost).

Switch to vectorcall.

Fixes #6666