Fix definition in pyx not overriding declaration in pxd
If a static method is declared in a `.pxd` file, we're not overriding `self.entries[name]`
with the definition in the `.pyx` file.

The difference in behavior with 3.0.x was changed by #3235.

This issue surfaces as there being two distinct entries in `cfunc_entries`, causing the one
declared by the `.pxd` to surface this error when checking that the `.pyx` implements all
methods declared in the `.pxd`:

```
=== Got: ===
13:15: C method 'meth' is declared but not defined
```