Speed up method calls with PyObject_VectorcallMethod (GH-6747)
Replaces most of our specialized "PyMethodCallNode" code with
"PyObject_VectorcallMethod". This avoids "__Pyx_PyMethod_New" (which is slow
in the Limited API specifically) and essentially lets Python do all the
unpacking stuff that we don't have access to.

This is something like a 35% speed up on the "bm_richards_pyclass" code in the Limited API.