Add warnings on likely misuse of critical_section
As a decorator of a Py-class function should be visible by default
because I can think of almost no good reasons for it.

Python attribute access is a lower-level warning because it seems
like it sometimes makes sense - while that operation isn't protected
and will likely break the critical section it can be reasonable
as part of a larger critical section.