Simplify __Pyx_PyType_Ready for single inheritance case
The reason I'm worried about this is that I don't think it's
possible to turn the GC on and off in a thread-safe manner.
This PR doesn't fix that problem, however if does avoid
having to interact with the GC in most cases so it makes it
less likely that we'll hit any issues.

(I have what may be a better fix in progress, but I thought I'd pick the
simple part of this out since it can stand alone, while the better
fix may need more thought)