Avoid turning GC off for static types
My concern is that it isn't possible to turn the GC on and off
in a thread-safe way, because we don't know what other threads
(including but not limited to other Cython module imports)
are doing with it at the same time.

This solves the problem in a different way. It creates a dummy
metatype that reports that it isn't gc-able, and
then temporarily swaps in that metatype.