Allow construction of non-moveable C++ classes
... with the cpp_locals directive.

This is mainly with a view to some of the C++ thread
synchronization classes (like std::latch), which are neither
default constructable nor-moveable, so fits in badly
with Cython's code-generation.

They can be constructed using "emplace" instead and so this is
what I do.

(I have a longer-term view to make this work outside of cpp_locals,
which I think is still viable for simple cases. However, that
needs a bit more thought about how to make it work).