Namespace-only named cpdef enums (GH-6382)
No longer copy the items of names cpdef enums to the module namespace.
Keep this behaviour for the items of anonymous enums and public/api enums.

The "module namespace polution" has proven to be a reason for users
not to adopt `cpdef enums` at all, so this change removes that issue.

Closes https://github.com/cython/cython/issues/4571