Fix some issues with C++ templated function overloading
Don't mark extern C++ methods as "inherited" - this isn't correct
and I think was probably only there to make the compiler
ignore some signature mismatches.

Fix an issue where methods returning a pointer wouldn't be marked
as a const method correctly (this was meaning that some const
accessors for C++ allocator) were being seen as having the same
signature, generating a warning.

Treat const methods/non-const methods as having a different
signature. This means that both are declared as alternatives.

Treat methods with different numbers of template parameters
as having different signatures (so that they appear in the list
of alternative).

Add some logic to IndexNode to look at the list of alternatives
and select one with a suitable number of template arguments when
there's multiple options.