Fix __module__ attribute for shared types
It can't be a descriptor because that messes up
`type(func).__module__`. I decided not to put it in `__dict__`
because that involves instantiating `__dict__` for every instance
(which we've previously decided not to do) and also deciding on
the behaviour when `__dict__` is replaced.

Therefore the solution seems to be to add some special casing
to tp_getattro and tp_setattro.

Fixes #6841.