Split exception_value into python repr and C repr
Python repr is used for working out if functions are compatible
(e.g. the same signature with `except -1` and `except -1.0`
continues to be compatible).

The C repr (as a string) is used for code generation, which
lets you use things like infinity as an exception value.

Add a warning on NaN exceptions because they won't work (and
have never worked). And they're also quite a tempting thing to
do.