Split exception_value into python repr and C repr (GH-6903)
Python values are used for working out if functions are compatible
(e.g. the same signature with `except -1` and `except -1.0` continues to be compatible).

The C repr (as a string) is used for code generation, which lets you use
things like infinity as an exception value.

Add special handling for (potential) NaN exception values because they do not compare equal.
And they're also quite a tempting thing to do.

Fixes https://github.com/cython/cython/issues/6900