Close a freethreading loophole on bytearray loops
Although we did already account for the possibility of the object
changing size during, it wasn't handled in a thread-safe way.

There's a slight discrepency with Python here (and for list
iteration too) - in Python I think the iteration should always
just end. In Cython, we may end with an `IndexError` if
the sizes changes at the wrong time. I can live with that.