Add Python 3.13t to benchmarks
There's definitely some pessimizations both in our code (and I
think in the interpreter) so it'd be good to see how much worse we
are.

It's a bit ambiguous if we should use 3.13t or 3.14t, but I
don't think it really matters for the sake of getting a quick
glance. When 3.14 is out properly we should definitely update
it though.