Avoid pessimizing sequence checks on 'range' (#6899)
It's possible that we might want to generalize this rule, rather than
leave it as a very specific special-case. However: `range` is going to
be a very common function to be looping over and we know the result
isn't a tuple or list. So let's not tell that compiler that it is.