Allow any Python number type as input to PEP-604 'x: float | None' type function arguments (GH-6808)
Convert the value to Python float on the way in.

The reasoning is that Python typing and static type checkers explicitly allow `int` etc. as input.

Incidentally, Cython also allows arbitrary Python numbers as input for
`x: float` and `x: cython.double` arguments (converting them to C double),
but not for `float | None` or `Optional[float]` that require actual Python objects.

Closes https://github.com/cython/cython/issues/5750