see if not hashing the directory helps ccache
For the benchmarks it looks like it's reporting a 0% cache hit rate, which doesn't seem worthwhile. 

In principle the directory affects it I think, and the benchmark runner uses a temp directory, so try to exclude that. 

I'm not quite sure how I know it's worked though - possibly just run the benchmarks twice and see what happens?