Fix error handling in nogil indexing of bytearray and str (#6947)
Cython has an existing feature where it allows indexing of bytearray and
str without the GIL. However it doesn't restore the GIL to report errors
and so will crash given an out of bounds error. Fix and test that.

Notes:
1. That Cython doesn't allow nogil indexing of bytes. I don't attempt to
change that here.
2. The whole thing feels *extremely* dubious in the Limited API but
practically you get away with it (so I'm ignoring this...)