Mark utility function static to avoid ODR violations (#6957)
PR #6882 introduces a new method `__pyx_CommonTypesMetaclass_get_module`
in `CommonStructures.c`. Unlike other functions in this file, it is not
marked `static`, which causes one-definition-rule violations when trying
to compile multiple `.pyx` files into a single extension module.

For context, this module structure is used in
[Cantera](https://github.com/Cantera/cantera). We build the combined
module
[here](https://github.com/Cantera/cantera/blob/b308592776130d82eaac4949283fc87c031cdc97/interfaces/cython/SConscript#L46-L81),
and then use a custom "finder" to load the submodules as defined
[here](https://github.com/Cantera/cantera/blob/main/interfaces/cython/cantera/_cantera.pyx).

Marking this function `static` restores the previous capability.