Use dict setdefault to handle setting to False
since it internally makes sure that nothing happens between
getting and settings. This means that all versions go through
a dict pointer and not just the CPython version.

Optimize __Pyx_PyDict_SetDefault a little for the Limited API
just because it's increasingly useful for this kind of
think so we may as well try to make it fast