Improve ccache and benchmarks
From testing it locally, it seems to have real difficulty with
them being compiled in temporary directories.

Setting the "basedir" to the temporary directory and telling it
not to hash paths seems to be sufficient for it to work for me
locally. Of course that doesn't guarrantee it'll work elsewhere...