Sets both .not_none and .or_none to False for annotated arguments. This makes the none flags for cpdef f(a: T) equivalent to those for cpdef(T a) which fixes possible cross module call segfaults.
Added run/cross_mod_nonecheck to check that something tests for
None before accessing an attribue on a C pointer -- doesn't
segfault because it's returning part of the memory for the None
instance.

Note that not all tests succeed. If there is interest in merging
this, I can fixup the other tests and / or limit the scope of this.