Store module strings as ompressed concatenation (GH-6971)
Rewrite the string storage to use compressed concatenated strings and an index to split them into the final user strings.
Allow users to chose between zlib, bzip2 and (Py3.14-only) zstd compression.
With zlib, the gain for the benchmark modules is 2-8%, but can be slightly negative for (small) modules with very little string content due to the increased code.