Add Limited API macros into api/public headers
This is mainly just preparation for sorting out pymutex in
public headers (since I think we'll want to use the Limited API
macros to support it).

However - it also simplifies the import and export code slightly
and is one less thing to remember to do differently in this
header code.

Added a test since it seemed pretty untested.