Shortcut some freethreading indexing
Essentially, in a lot of cases a refcount of 1 tells us that
no other thread has access to a variable and thus we can
directly use the fast macro indexing. A micro-benchmark
suggests this is twice as fast where it hits, and also
that it doesn't cause much measureable difference where it
misses.