Don't put the DEL character in character strings (#7024)
Fixes #6982

I can find conflicting sources on whether it's printable or not, but it
doesn't look great in my editor and it doesn't appear in the C language
standard of what's allowed to appear in a character constant
(https://en.cppreference.com/w/c/language/charset.html).

I don't think it's causing problems with real compilers but equally I
don't think it hurts to escape it instead.