Add thread-safety check to async-gen initialization
There may not be a huge value in this (if nothing else, CPython
doesn't have this check, although I expect they eventually will)
but guard against initialization of an async generator while
it's already in progress.

I've chosen to raise an exception but other reasonable options
would be to skip it (and assume the order doesn't matter)
or to block until the first thread finishes (which is a little
harder to arrange).  Given this is testing fairly silly
user-code I don't think this matters.

This only handles the initialization part. The rest of the
file remains as is.