Improve heavily contended PyThreadType lock (#6999)
It should be safe to block for the GIL while holding the lock provided
we never do it the other way around (which we don't because we control
access to the lock). That probably makes things worse if you're mixing
locking with and without the GIL, but it does at least mean that you
eventually succeed.

Add a manual benchmark that can be used to measure it. I don't think it
makes sense as an automatic benchmark because you're interested in both
wall-clock time and CPU time, but it was useful while investigating
this.

I also note that in some cases (very contended, with both timeouts set
to 0) the PyMutex never seems to finish. This isn't something we can
really control but arguably it might be worth documenting.

Fixes https://github.com/cython/cython/issues/7023 and the comment
https://github.com/cython/cython/pull/6579#discussion_r2176556402