Add __Pyx_PyDict_NextRef (and related functions)
Essentially a version of PyDict_Next that returns on owned
reference for keys/values. On platforms with CYTHON_AVOID_BORROWED_REFS
it doesn't ever go through a borrowed reference which I think
should help because it reduces the amount of objects that are
ambiguous if they can be cleaned up.