Make "PyImport_AddModuleRef" somewhat thread safe
It turns out that it isn't (for now). It's isn't generally
possible to make it thread safe given that people can replace the
module dict with arbitrary mappings, but we can at least handle the
common sensible case where it's a dict.

I'm mostly concerned with the shared utility module here.