Improve GraalPy support (#7074)
I'm adding a few things that should improve compatibility with
GraalPy. I was testing with pre-release build of upcoming GraalPy 25.0
(implements Python 3.12).
- Use datetime macros instead of struct access
- Set `CYTHON_AVOID_THREAD_UNSAFE_BORROWED_REFS` to `0`. This macro
seems to lead to code paths that assume Python 3.13.
- Use new GraalPy public functions that are prefixed with `GraalPy`. We
had to do a rename to prevent future name collisions with CPython (i.e.
imagine if they add `PySlice_Start` with different refcounting
semantics) and to make it clearer which APIs are GraalPy-specific.
- Add partial support for `array`. Currently only implements the access
to `data` and the resize functions that are used by scikit-learn