Replace module state lookup with hash table
When using multi-phase init and module state, use a hash
table instead of bisection into an array. This should hopefully
perform better on average (especially when there are
many modules available).

Special case the 0th module so this one's always available
quickly.

Make it so the implementation can be tested independently
on subinterpreters/module state.