Update C/C++ py_safe locking loops...
to match those in 3f00d92 (i.e. be a bit more usable in
heavily contended cases).

This does make it slightly easier to deadlock them if someone
fails to use the py_safe mechanism somewhere when they
genuinely do have the GIL. However, I'm convinced
not to worry about that just because the GIL can be arbitrarily
released and reacquired anyway by Python-like code, and
so we can't guard against this deadlock.