Optimise fused function dispatch (GH-7098)
Replaces the previous index with a a lazily initialised dict cache.

The new implementation speeds up the unambiguously matching cases into a
single dict lookup and makes the first lookup and the mismatch cases
expensive. This seems the right tradeoff.

We could potentially also cache signature mismatches, but then those
could end up unbounded and bloat the cache, leading to potentially high
memory usage and risking to open up a DOS attack vector if call
signatures are somehow untrusted.

We could try to find a middle ground by only caching call signature
mismatches that use "expected" argument types (i.e. those existing in
any of the fused function signatures), but that would only be to speed
up the exception raising. If we consider that an actual error, then
users can probably live with a slow response.