PEP 701 expanded fstring syntax (#7061)
Closes #5452

As described in PEP, a lot of the logic is moved into the tokenizer. So
we have a specialized f-fstring states covering the string part of the
fstring. When it hits a `{expr}` it goes into an `FSTRING_EXPR_PRESCAN`
which has enough logic to extract the expression into a string (to allow
self-documenting fstrings to work) but returns to the `fstring` state on
a suitable `:` or `}`.

I haven't updated `strip_string_literals` - the main issue I can see
with that is that it fails to identify the format specifier as string
literals

```
f'{a:cimport somemod}'
```

That's unlikely to be a useful format specifier, but would get
identified as a `cimport`. I'm ignoring this for now.

PR includes https://github.com/cython/cython/pull/7057 and the branch
https://github.com/cython/cython/compare/pep701_fstrings.