Reduce redundancy in fused type dispatch functions (GH-7100)
Move the type mappers into separate utility functions that are deduplicated and reused.
Collect the mapped argument types in local variables instead of a Python list.
This saves both space and time.

Also remove the Tempita usage for the simple code generation of fused types and replaces it with faster and simpler f-strings..