Fix dealloc from class inherited from a pxd file (#7143)
If we don't manage to look up the base classes dealloc function then we
really need to implement our own.

With static types it's fine as is - the base class dealloc function is
automatically inherited. With heap types, if we leave the dealloc slot
as 0, then it sets the
dealloc function to an unsuitable default (which can end up causing
infinite loops). Therefore we do need to provide an implementation.