Disable parallelism in `cythonize` cmd on `spawn`
A similar problem was fixed in #3263 which is what this patch is
porting to another module. That fixed the Python-land API while this
change aims to also fix its command-line counterpart.

The symptoms are Python 3.8+, macOS and the underlying compiler
invocations failing. If these are met, `cythonize` would hang before
this commit.

Fixes #3973.

Refs:
* #3262
* #7146
* https://github.com/ansible/pylibssh/issues/769