Support opaque PyObject in cyfunction/fusedfunction
This is planning ahead a bit for when we want to support
Limited API + freethreading builds (which I'm strongly assuming
is coming in Python 3.15).  In this case PyObject will become
opaque because it's different between freethreaded and regular
builds and so Cython will need to handle most inheritance
without including PyObject into the type struct.

This PR only covers cyfunction/fusedfunction and not the rest
of the shared types, mostly as a proof of concept. I've also
not done it outside the Limited API, just because the regular
API does the weird "semi-inheritance from PyCFunction" trick.
For regular cdef classes there might be some value in giving
users individual control.