Fix CI for Python 3.14.0 (GH-7197)
With CPython turning its π series from rc into final, the `coverage_*`
end-to-end tests started failing.

It was confirmed that the CI passes with Python 3.14.0rc3 and older. But
the final release makes said test crash. This is because it never ran on
release candidate of 3.14 — the test-runner skips coverage tests on
pre-releases:
https://github.com/cython/cython/blob/5b14ba8/runtests.py#L248. Perhaps,
that check should allow release candidates too.

The reason for the coverage tests failure is that coveragepy uses the
`sysmon` engine under Python 3.14 which ends up not printing coverage
for the Cython module and the test does not find the filename in the
output. But the test also disables sysmon while compiling. The included
fix explicitly makes coveragepy use the former default of `ctrace` to be
consistent with the Cython-side setting.

It's also failing on `pip install` locally, because PEP 517 build
backend is isolated and the context `setup.py` (having said import) is
running in does not have access to the "outer" Cython copy. It is
unclear why this works in older runtimes.

This patch forces the PEP 517 build to happen with isolation turned off.