Replace __Pyx_PyObject_GetMethod with vectorcall (#7214)
Where we support vectorcall use it instead of __Pyx_PyObject_GetMethod.
This is partly to avoid having to think about getting the dict
atomically, but also because it should be the same or faster (at least
in the non-error case).

Partial fix for #7200 (there's another unrelated usage that I'd like to
sort out).