Implement "unified" lookup mechanism for module state
The idea is that all lookups for the module state go through
`__PYX_GET_MODULE_STATE()` and the code generation no longer
needs to care about what type of lookup it might be (so doesn't
need to know anything about the scope).

At the start of each function we create a temporary definition
for __PYX_GET_MODULE_STATE() using one of a number of
mechanisms depending on what information is available in the scope.