Run abi3audit on ABI3 wheels (#7267)
In addition to a little bit of cleanup on the toml file mainly to reduce
duplication.

The cibuildwheel documentation recommends running it as part of "repair
wheel" rather than the tests. I didn't do that mainly because I could
specify test requirements but not a repair wheel requirement. The
consequence is that it gets run multiple times.

I also dropped the override on the "armv7l" CFLAGS so it didn't get a
few flags related to x86_64 (which didn't cause problems but seemed a
bit silly)