Fix an off-by-one error when padding Unicode characters with the ':c' formatting type (GH-7299)
Also fix an exception when trying to format lone surrogates this way.

And, along the way, I finally split the `:c` formatting from the other
integer formatting functions because it really is a different thing. The
decision is currently taken from a C macro because the (cached) type
conversion function cannot currently depend on the format string but
only on the Cython type of the formatted C value. The C compiler should
be happy to discard the unused function respectively.

Closes https://github.com/cython/cython/issues/7298