Fix crash when assigning nested structs from a list of structs (GH-7309)
Example of failing code (where NestedStruct is a struct consisting of a struct and a union):
`cdef NestedStruct ns = [Point(1.0, 2.0, 42), int_or_float(i=1)]`

The crash is due to the special item handling in ListNode. The type analysis converts the struct type call to a DictNode, but then ListNode takes the original items (i.e. the call) for the assignment, assuming that they are probably a better source that is less impacted by list item object conversions. This is generally true, except if the coercion does not happen through a CoerceToPyTypeNode but by actually replacing the node completely, as struct/union type calls do.

Closes https://github.com/cython/cython/issues/7308