Add a "threads" tag to tests
just as an easy way of selectively running tests that involve
multiple threads. Note that I've only added it to tests that
actually run, and not bothered with a few related error tests.

Also add a couple of missing openmp tags.

Will be part of #7334 but submitted separately because it's
pretty simple and independent.