Clean up checks for tp_dictoffset
It is inherited which means that we don't need to redefine
it for classes that have dict via inheritance.

Instead we should allow "this class has a dictoffset" or
"one of its first bases has a dictoffset".

Factor out some common limited API/type slots code to
find the dictoffset of a type.

This change is part of the bigger changes to support
limited API in Python 3.15(t)+ - in this version the base
type won't be the first member of of the object struct,
therefore meaning that we can't easily get offsetof the
dictoffset. It's taken out because it's standalone so
it probably simplifies the review of that change in the future.