Fix const casts for C string literals (GH-7346)
Turns out that "type == c_const_uchar_ptr_type" does not exclude non-const types.
And we probably won't find a `char * const` anywhere but would rather be
interested in `const char *` for a const cast.

I'll mark 3.2 as the target version because I'm not sure we should
change it in 3.1 as well. Difficult to say what it might break.