Allocate internal C arrays for C array literals in an appropriate place (GH-7339)
* constant C arrays: `static const` C arrays in place
* module global: static C arrays in the module init function
* inside generators: non-const arrays stored in the generator closure
* otherwise: non-const local temp array variables on the stack

The constant case specifically applies to for-loop iteration over strings etc.
It avoids the need to store them away in generator closures and generally wasting runtime work for them.

Factors a dedicated `CArrayNode` out of the `ListNode` for the "list as C array" case,
to avoid complicating `ListNode` for this.  I kept `ListNode` as a base class mostly just
in case we need some of its functionality later on, but I'm not aware of anything specific
that is needed.  It could probably just be a `SequenceNode`.  We should be looking
at `is_sequence_constructor` rather than specific `ListNode` / `TupleNode` classes
wherever it's needed.

Also enables coercion from a Unicode string literal to an integer C array of character values
(not just `Py_UCS4[]` but also e.g. `short[]` for more compact storage of BMP Unicode characters).
I think that's a reasonable thing to have.  It doesn't do any compile time value checking,
but the C compiler would probably warn.

Closes https://github.com/cython/cython/issues/7323