Fix maximum number of ProcessPool workers (#7385)
Set the default for the maximum number of workers to `None` which lets
the executor decide. Practically that makes it equal to the number of
cores or the internal limit (whichever is lower). That's lower than our
1.5 * cores (which I imagine is to leave room for a bit of IO) but at
least will never hit the limit.

The other usage of ProcessPoolExecutor (in Dependencies.py) always takes
a number rather than guessing a default, so should be safe.

Fixes #7384