Don't abi3audit cython-inline files
test_coroutines_pep492 seems unique in that it builds the
inline files inplace. The other way to fix this would be to
remove

```
lib_dir=os.path.dirname(__file__)
```

from that. I haven't dne this because I think building in
place is probably a good thing.

Ideally of course we'd propagate the cflags to cython_inline.
I've done this for the srctree tests, but the inline ones are
more adhoc so that's probably harder.