Add a C macro to configure "Py_mod_gil" at C compile time (GH-7404)
I found a use case for setting `Py_mod_gil` differently in Py3.13 and Py3.14,
but regenerating the whole C file just to switch between them is rather wasteful.
So here's a new C macro `CYTHON_FREETHREADING_COMPATIBLE` that allows
defining the setting from the outside.

Having a directive is still useful because we don't want to require the
C macro to be set. It should usually have a fixed value for a given module.