Make fused function signatures a non-writeable dict-proxy (#7386)
Slightly related to #7383 (I think that classobj should be set once in
advance and that's easier to reason about if people can't change the
contents of `__signatures__`; however it's only fairly losely related).

I've only changed the user-visible attribute - I did consider changing
the internal variable but then we can't use optimized access when
dispatching (which seems worse).

My view is this should go into 3.3 and hopefully there's plenty of time
for anyone who is doing silly things messing with `__signatures__` to
complain if needed.