Refactor of vtab access
This is part of the Stable ABI/opaque objects changes. I've
picked it out in the hope of keeping it fairly simple to review.

In the opaque objects world it doesn't really work
to lookup multiple levels via `p->base.base.base` because the
object struct we have doesn't contain `base`.  This means that
it's better to access the relevant base by casting (which will
be changed to `PyObject_GetTypeData` in the opaque objects
version).

The result is that it's better to store the type in the object
heirarchy that contains the vtab slot rather than a cname
describing how to get to it.