Optimise object comparisons against single Unicode character literals (GH-7418)
This is used a lot in our parser.

This turned out a bit tricky because we cannot easily change the signature
of optimised C comparison functions, and we also cannot reverse the order
of the operands here due to the way cascaded comparisons work.

Thus, I chose to generate a simple macro for each character that we
compare against, named after the character code, and then let the C
preprocessor inject the integer of the character code in its
replacement. This might appear a bit wasteful, but usually, the number
of different characters that user code compares against is rather limited.