Add Limited API 3.15 test (#7407)
Limited API for Python 3.15 is likely to be different enough that it's
worth starting to test it now (i.e. it will probably involve opaque
objects).

There's some danger that CPython will get ahead of us in making opaque
objects compulsory and so the test will start failing - I think that's
OK and we can just turn it off.

I've assumed that abi3audit won't work for the moment (they're unlikely
to be strictly up-to-date with the latest changes).