Avoid immortalizing anything shared on freethreading
Py_SET_REFCNT isn't thread-safe so there's the
risk of races if another thread is accessing it at the same
time. This will mostly apply to things that are already
immortal so I don't think we'll lose much from the
check.

On Python 3.13t we have to use a slightly weaker check
than the full "IsUniquelyReferenced" - I think this is fine
because we know that our "1" reference is not yet accessible
anywhere external, however I've still chosen to use the full
check where available.